agns_core/pyproject.toml
[project]
name = "agns-core"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "orjson",
    "pydantic",
    "uvloop",
    "PyYAML"
]

[tool.uvicorn]
factory = true
